alx_travel_app/requirements.txt
django
djangorestframework
django-cors-headers
celery
rabbitmq
drf-yasg
django-environ
mysqlclient
faker